registro-notas/datastore-service/requirements.txt
google-cloud-datastore
google-cloud-pubsub
tornado
